completed version 1.0 of partial retrieval caching, still need to add documentation
